feature: Introduces eth backed hotel with url field on-chain abstraction
